requirements.txt
flask
firebase-admin
google-generativeai==0.4.1
python-dotenv
flask-cors
